# docs/swagger.yaml
definitions:
  be-dashboard-nba_internal_presentation_presenter.Pagination:
    properties:
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total_items:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.Pagination'
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_menu.MenuParent:
    properties:
      group:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.CreateMenuPermissionRequest:
    properties:
      action_name:
        maxLength: 50
        minLength: 1
        type: string
      code:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - action_name
    - code
    type: object
  be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.ReadMenuPermissionResponse:
    properties:
      action_name:
        type: string
      code:
        type: string
      id:
        type: integer
      menu_id:
        type: integer
    type: object
  be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.UpdateMenuPermissionRequest:
    properties:
      action_name:
        maxLength: 50
        minLength: 1
        type: string
      code:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - action_name
    - code
    type: object
...
info:
  contact: {}
  description: Template API for Golang project
  title: Base APP Go
  version: "1.0"
paths:
...
  /menus/{menu_id}/permissions:
    get:
      description: Retrieves a paginated and searchable list of permissions (e.g.,
        read, create) for a specific menu ID. Requires Bearer token.
      parameters:
      - description: ID of the Menu
        in: path
        name: menu_id
        required: true
        type: integer
      - description: Search term for action_name
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Sort order (e.g., action_name ASC)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get menu permission list
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.ReadMenuPermissionResponse'
                  type: array
              type: object
        "400":
          description: Bad Request (Invalid Menu ID)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Not Found (Menu ID not found)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error (Query parsing or usecase failure)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Get Menu Permission List
      tags:
      - Menu Permission
    post:
      consumes:
      - application/json
      description: Adds a new permission (e.g., 'read', 'create') to a specific menu
        by its ID. Requires Bearer token.
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: integer
      - description: Permission data to create
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.CreateMenuPermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully create menu permission
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid Menu ID, Invalid JSON, or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Create Menu Permission
      tags:
      - Menu Permission
  /menus/{menu_id}/permissions/{menu_permission_id}:
    delete:
      description: Deletes a specific permission from a menu by its unique permission
        ID. Requires Bearer token.
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: integer
      - description: Menu Permission ID
        in: path
        name: menu_permission_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete menu permission
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid ID)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Not Found (Permission ID not found)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Delete Menu Permission
      tags:
      - Menu Permission
    get:
      description: Fetches the details of a single menu permission by its unique ID.
        Requires Bearer token.
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: integer
      - description: Menu Permission ID
        in: path
        name: menu_permission_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get menu permission detail
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
            - properties:
                data:
                  $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.ReadMenuPermissionResponse'
              type: object
        "400":
          description: Bad Request (Invalid ID)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Not Found (Permission ID not found)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Get Menu Permission Detail
      tags:
      - Menu Permission
    put:
      consumes:
      - application/json
      description: Updates an existing menu permission (e.g., 'code', 'action_name')
        by its unique ID. Requires Bearer token.
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: integer
      - description: Menu Permission ID
        in: path
        name: menu_permission_id
        required: true
        type: integer
      - description: Permission data to update
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.UpdateMenuPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update menu permission
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid ID, Invalid JSON, or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Not Found (Permission ID not found)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Update Menu Permission
      tags:
      - Menu Permission
  /menus/parent:
    get:
      description: Retrieves a list of all root-level menus (menus without a parent),
        typically for use in a dropdown. Requires Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get parent menu
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu.MenuParent'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed get parent menu
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Get Parent Menus
      tags:
      - Menu
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your token}" into the field below.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"