# internal/presentation/handler/menu/menu-permission/delete_menu_permission_handlers.go
package handlers

import (
	"be-dashboard-nba/constant"
	contract "be-dashboard-nba/internal/domain/contract/usecase"
	"be-dashboard-nba/internal/presentation/presenter"
	"errors"
	"net/http"
	"strconv"

	"github.com/gofiber/fiber/v2"
)

// DeleteMenuPermission godoc
// @Summary      Delete Menu Permission
// @Description  Deletes a specific permission from a menu by its unique permission ID. Requires Bearer token.
// @Tags         Menu Permission
// @Produce      json
// @Param        menu_id            path      int      true "Menu ID"
// @Param        menu_permission_id path      int      true "Menu Permission ID"
// @Success      200 {object}  presenter.ResponsePayloadData "Successfully delete menu permission"
// @Failure      400 {object}  presenter.ResponsePayloadMessage "Bad Request (Invalid ID)"
// @Failure      401 {object}  presenter.ResponsePayloadMessage "Unauthorized"
// @Failure      404 {object}  presenter.ResponsePayloadMessage "Not Found (Permission ID not found)"
// @Failure      500 {object}  presenter.ResponsePayloadMessage "Internal Server Error"
// @Security     BearerAuth
// @Router       /menus/{menu_id}/permissions/{menu_permission_id} [delete]
func DeleteMenuPermission(svc contract.MenuPermissionUseCase) fiber.Handler {
	return func(c *fiber.Ctx) (err error) {
		menuPermissionIDParams := c.Params("menu_permission_id")
		menuPermisionID, err := strconv.Atoi(menuPermissionIDParams)
		if err != nil {
			return presenter.ResponseMessage(c,
				presenter.ResponsePayloadMessage{
					Code:    http.StatusBadRequest,
					Message: "Failed parse params",
				})

		}

		err = svc.DeleteMenuPermissionUseCase(c.UserContext(), menuPermisionID)
		if err != nil {
			if errors.Is(err, constant.ErrMenuPermissionIdNotFound) {
				return presenter.ResponseMessage(c,
					presenter.ResponsePayloadMessage{
						Code:    http.StatusNotFound,
						Message: constant.ErrMenuPermissionIdNotFound.Message,
					})
			} else {

				return presenter.ResponseMessage(c,
					presenter.ResponsePayloadMessage{
						Code:    http.StatusInternalServerError,
						Message: "Failed update menu permission",
					})
			}
		}
		return presenter.ResponseData(c, presenter.ResponsePayloadData{
			Code:    http.StatusOK,
			Data:    nil,
			Message: "Successfully delete menu permission",
		})
	}
}